+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # 24/02/2023
 
 - Added Sentry error tracking
-- Delete a session
+- Added ability to delete a session